Create generator-generic-ossf-slsa3-publish.yml
Veyron Pax the founder of Gr00nd is a multidisciplinary artist Having him the creative director with design background, the founders realized how difficult it was to find the perfect assets for a design. Gr00nd was created to improve and change accessibility of design tools and in general different art forms.  
He defines value in the originality of a vision thus having a full access to platform, instrument and materials is what supports the artists and everybody else to benefit from.
His ultimate goal is to make believe of a safe supported and free environment for artists to be able to use their most potential for creativity.